Fix multithread bug: Reset _run_args, etc.
Problem: Imagine executing Script.run for a Script that returns
$1 (i.e., its first parameter). $1 is a substring of _run_args.
_run_args is a member of Script. Script.run is multithread-safe
(it's an "exclusive" handler, so at most one Script.run can run
at a time). But we didn't reset _run_args at the end. So: the
return value of Script.run shared memory with Script::_run_args.
The return value is manipulated by clickfs, and eventually
dereferenced and freed, OUTSIDE the configuration lock. But
Script::_run_args is reset the next time Script runs, which
might be concurrent! Crash.

Derrick Pallas reported the problem and helped narrow it down.