ToDevice: remove deadlock
ToDevice calls Element::pull with the tx_lock spinlock held.  Since Elements
might do something that schedules, this can lead to deadlock.

I noticed this because in our click config, an Element schedules a timer in
the pull path, which should be allowed but which also acquires a mutex in
order to update the TimerSet.

The fix is to release the lock around input(0).pull() and then trylock.  On
fail, we bail out, skipping all of the code that might require the lock to
be held.  There is relatively little performance hit in this solution.

An alternative fix would be to create a Deque<Pair<Packet*,click_jiffies_t>>
and pre-fill it with _burst packets before acquiring the lock.  I tried this
and performance suffered.  It might work if the netdev could tell us roughly
how many Packets it would be willing to TX.  Without that information, we
add an extra, unnecessary layer of buffering.