Reference platform header for in6_addr struct
Prior to this commit, click defined it's own struct for an IPv6
address. This change modifies to be consistent with the IPv3
implementation in that it pulls the standard 'struct in6_addr'
definition from the platform include files.  This also give access to
some standard IPv6 related macros/constants. Updated files in this
commit were just to reflect the name change from 'click_in6_addr' to
'in6_addr'. References to s_addr16 and s_addr32 were removed as these
are not universally defined on all platforms.

This passes all click tests.