Update support for embedded IPv4 Addresses
This change updates core IPv6 addressing related components to the
latest IPv6 Addressing Architecuture (RFC 4291). Specifically, it
splits support for IPv4-Compatible addresses, deprecated in RFC
4291-2.5.5.1, and augments support for IPv4-Mapped addresses.
IPv4-Mapped address are never transmitted over the wire. Instead, they
are used internally to represent v4 addresses in a single data
structure. The IPv4-Mapped form is the default for any IPv4 Address
assigned to the IP6Address class. The basic_parse routine was modified
to also accept IPv4 AddressInfo mnemonics of the form "0::ffff:<tag>".

Simplified the ip4_address() function to simply return either the v4
address, if it's v4-mapped, or the zero address.  The result can be
used directly in a bool test without the need for another variable.