click: add support for debug tracing packets
Allow packets to be marked for tracing. With each push or pull of a
packet marked for tracing, Click chatters a line indicating the source
and destination elements and ports.  Also, Click chatters a line when a
traced packet is killed.  The output looks like this (from a testie):

  traced_packet: trace_push [0] -> [0] RoundRobinSwitch@3
  traced_packet: RoundRobinSwitch@3 [0] -> [0] isw
  traced_packet: isw [0] -> [0] Discard@5
  traced_packet: killed

Arbitrary elements can mark packets for tracing. However, the typical
way to mark packets is through the Trace element added in the next
commit.

I opted for a really simple implementation: a static Packet member that
stores a pointer to the traced packet. (Note that non-static Packet
members are not allowed in the Linux case.) This pointer is never
dereferenced - it just lets pushes and pulls identify a traced packet.
The implementation is not entirely thread safe: in particular, on
platforms where pointer assignment is not atomic, the pointer may end up
containing garbage. However, since the pointer is only used for
comparison and this is just a debugging feature, this should be OK.  The
static packet pointer is cleared whenever a traced packet is killed or
leaves Click. I believe linuxmodule's ToDevice and ToHost are the only
places a packet leaves Click without being killed.

With this implementation, only one packet can be traced at a time. In
particular, this means that normally you can't start tracing a packet if
another packet is still being traced. This behaviour can be changed
using the 'preempt' argument to set_traced().

Asa Zernik added code for printing the output port for push() and the
input port for pull(). Jignesh Patel added code for preempting tracing.

Reviewed-by: asa@meraki.com
Signed-off-by: Patrick Verkaik <patrick@meraki.net>