userlevel click packet: don't use kill() to unreference a packet
In Linux click, Packet::kill() is called on a packet just once -- when
the user of the packet discards the packet -- at which time the packet
is recycled. Sharing data among packets to implement clone() is handled
outside of Packet, and killing clones of a packet does not affect the
cloned packet.

In userlevel (and other non-Linux) click, packets share data by having a
clone keep a reference to the cloned packet. When the clone is killed,
Packet internally calls kill() on the referenced packet to decrement a
reference count and (if the reference count drops to 0) recycle the
packet.  This means that a packet's kill() may be called several times:
once when the packet's user discards the packet and once each time a
clone is killed.

Unfortunately, the userlevel click behaviour breaks any code in kill()
that needs to run only when a packet is discarded by its user, in
particular packet tracing code.  For packet tracing, a packet may get
untraced prematurely if a clone of the packet is killed before the
packet itself.  This change fixes the issue by factoring out
dec_use_count() and having Packet use that instead of kill() to drop a
reference to a packet.

Reviewed-by: asa@meraki.com
Signed-off-by: Patrick Verkaik <patrick@meraki.net>