clickfs/llrpc: make non-safe calls stop router threads
Previously, "safe" LLRPCs would take the read config lock and normal LLRPCs
would take the write config lock.  However, normal LLRPCs would not block
the router, meaning that shared data structures still needed to be protected
inside Elements.  This becomes extremely complicated if the Element needs to
push during the LLRPC or if a shared data structure needs to be accessed in
both atomic and non-atomic contexts.

This change makes normal LLRPCs also block the router.  The alternative was
to add an additional "exclusive" class of LLRPCs but it is hard to think of
an case where this call needs to be the only LLRPC+handler running but does
not also need to stop the router.

The only exception here is CALL_HANDLER, which will block based on whether
the handler itself is marked as exclusive or not.

Signed-off-by: Derrick Pallas <pallas@meraki.com>