skbmgr: fix skb_recycle call
In older kernels, skb_recycle would return the SKB if it was recycled and
NULL if it was not recyclable.  Later, this function became void and assumed
that the caller had verified that skb_is_recycleable is true.  This is
encapsulated in skb_recycle_check.  In newer kernels, both have been
removed.

This commit changes click to use skb_recycle_check instead.

Signed-off-by: Derrick Pallas <pallas@meraki.com>

♲