skbmgr: only count SKB as recycle/freed if it is actually recycled
The count of recycle/allocated SKBs only goes up if an SKB is actually
recycled but the count for recycle/freed SKBs goes up even if the kernel
reports that the packet is actually not recycleable and it is not added to a
bucket.  This leads to a confusing report in /click/skbmgrinfo, where the
sum of bucket counts is small but the difference between recycle/freed and
recycle/allocated SKBs is very large.

This commit moves the recycle/freed increment so it is inside the "recycle
this SKB?" conditional.

Signed-off-by: Derrick Pallas <pallas@meraki.com>