linuxmodule: todevice: fix use after free when reallocating for more headroom
Earlier in queue_packet, a packet may have been reallocated to expand
the headroom. If a packet is both reallocated *and* the call to the xmit
function fails, this code queues up the freed packet instead of the new
skb.

This change is to always queue up skb1 instead of p, which may not be
valid anymore.