Userlevel version of fast(tcp|udp)flows and fastudpsrc(ip6)
FastTCPFlows, FastUDPFlows and FastUDPSourceIP6 are moved to tcpudp package and are compatible with userlevel and linuxmodule, with macros to wrap around the checksum generation which differs on linuxmodule or userlevel (FastUDPSourceIP6 is already compatible out of the box).

FastUDPSrc has a specific version for bsdmodule and doing the same would create a conflict so I made a userlevel-only version of FastUDPSrc. The 3 would need to be merged, but I don't have any bsd testbench.

Tested on linuxmodule and userlevel.

Signed-off-by: Tom Barbette <tom.barbette@ulg.ac.be>