Enable MP SC mode in linuxmodule FromDevice
As if using multiqueue with irqbalance (or manual pinning of each queue interrupt to multiple cores to share load) the enqueueing function will be called on the interrupt's core.

The producer use a CAS instruction to grab the head in the ring and advance it. As the consumer could read the ring[head] before the producer writes to it, the 0 value is used to specify a "writing pending" and the consumer spins until it reads a non-null value.

Signed-off-by: Tom Barbette <tom.barbette@ulg.ac.be>