Force PCAP to use TPACKET_V2 instead of TPACKET_V3.
As from libpcap 1.5.0 the default is to use TPACKET_V3 if available in the kernel. However TPACKET_V3 is catered for better packet capture and less packet drops, but is not suitable for capture and transmit scenarios as it introduces significant latency when forwarding packets in userlevel Click using FromDevice with METHOD PCAP. Setting pcap_set_immediate_mode forces PCAP to fall back to TPACKET_V2 in immediate mode.

See the following references for more information:
http://seclists.org/tcpdump/2014/q4/53
http://seclists.org/tcpdump/2014/q1/122

Resolves issue #164