Added DPDK Support
Provide 3 new elements to use Intel's DPDK framework (compatible from
version 1.7 to the current 2.1) to receive and output packets at very
high speed in Click. Configure using the --enable-dpdk to enable the
DPDK framework. [1] provides a comparison of throughput using the new
DPDK implementation and the Netmap one with the best configuration
achievable in a forwarding (without touching the packet) and a router
configuration.

FromDpdkDevice is used to receive packets and ToDpdkDevice to output
packets. DpdkInfo can be used to set some DPDK-related parameters, see
[2] and dpdkinfo.hh for more informations about them.

These I/O elements supports multiqueue, you can use the QUEUE
parameter to specify a queue index. If omitted, one queue will be used
per FromDpdkDevice/ToDpdkDevice

Difference from traditional ToDevice is that ToDpdkDevice only
supports PUSH mode, as DPDK does not allow select-like mechanism.
Anyway [3] shows that full-push mode leads to better performance most
of the time. However, under very low throughput this could induce
higher latency as the output device waits for a certain number of
packets (BURST parameter) before flushing. The TIMEOUT parameter
allows you to limit that time.

ToDpdkDevice is thread-safe, by using a vector of internal queues to
batch packets, and one timer per thread. It uses a spinlock when the
batch is ready to protect the physical device. Regarding the
FromDpdkDevice, as only one thread will spawn the Task, it does not
need such a lock or per-thread state. This mechanism is by far faster
than using a ThreadSafeQueue or an external locking element.

This patch is a stripped version of the DPDK support in FastClick
(mostly without batching and auto-thread assignment). User may want to
read [3] for implementation detail, better queue/burst usage and read
further about full-push mode.

For archive, we also tried a DPDK-pool mode, where we don't use Click
Packet object but a wrapper around DPDK's mbuf. While this saves a
little memory, it was slower as Click's Packet object are re-used and
provide a better memory locality.
Indeed, putting anotations in DPDK's mbuf header lead to two cache
miss per header instead of zero when Click can cycle through Packet
object, just changing the buffer pointers and reusing them. If people
are interested in this, they can try it in the code of [3].

[1] http://bit.ly/1Z2Eg4W - Performance of this new DPDK implementation
[2] http://www.dpdk.org - DPDK website
[3] http://hdl.handle.net/2268/181954 - FastClick paper