packet.hh: Fix shift_header_annotations in Linux >= 3.11
Prior to Linux 3.11, the locations of the various protocol layer
headers (MAC, network, transport) were stored in the skbuff as
pointers for 32-bit kernels and offsets for 64-bit kernels.  In Linux
3.11, this was changed so that they were always stored as 16-bit offsets
regardless of the kernel word size.  As a result the adjustments made by
Packet::shift_header_annotations were incorrect on 32-bit kernels
beginning with Linux 3.11

The actual manifestation of this problem was that the header annotations
were incorrect after an expensive_uniqueify.