dpdk: Add one travis test per DPDK version
Use a build matrix in travis to test all DPDK versions. This offer more parallelism in Travis as one concurrent job will be launched per framework : netmap, dpdk, or vanilla.

The exclude part is not very clean, but DPDK is not compatible with clang. This solution does not launch a job in travis for nothing. If I keep last method (a condition in the job itself to do nothing if $CC==clang), it would launch all jobs for clang, always exiting doing nearly nothing except installing packets. Not very clean either. Ultimately the better solution will come from Travis...