dpdk: Allow to give a PCI addr instead of port
The problem with port index is that it depends on the number of PCI
devices. Therefore if you want to target a specific NIC, you have to
know the complete PCI layout, and if a PCI device is added before yours,
all port index will be shifted... The PCI address are of course fixed.

This is done through adding a new DPDKDeviceArg parser, if it is given
a number, it will be used as a port index, if it is a PCI address it
will search through all DPDK-attached devices for a matching PCI address.

DPDKDevice is also changed a little-bit to be non-completely static and now
realy represents a DPDK Device. _port_id is now replaced by DPDKDevice _dev
in the From/ToDPDKDevice elements, and is passed as it to the PORT
argument. Click's nice argument parser do the rest  to call
DPDKDeviceArg::parser. The DPDKDevice now contains the port_id, and the
DevInfo.

Therefore the hashtable of all used DPDK devices is now a
port->DPDKDevice map, which makes more sense. It allows to be a more oo-style
than passing port_id everywhere.