linuxmodule: fix illegal expression in rbtree_latch
This header uses container_of to pick one of two branches in a binary tree,
and does so by passing an index of either 0 or 1 to __lt_from_rb.
Unfortunately, container_of needs to take a constant expression for the
third parameter; using this index breaks that.  Instead, make logic
explicit.