lib/packet: make expensive_uniqueify use linux skb interface
The original version of this function does the same job as pskb_expand_head,
except for the free_on_failure semantic, since we require an unshared SKB
as provided by skb_share_check.

If we want free_on_failure, skb_share_check will already take ownership of
the SKB.  If we do not want free_on_failure, then we can preserve the
reference to this with skb_clone and then Packet::kill() on success.

I wanted to refactor this code because I made some tweaks to net/core/skbuff
in Linux that break a couple of the assumptions here, specifically that
skb->head is kmalloc'd.  (Instead it comes from a pool.) Rather than pull
the alloc/free semantics into Packet as well, I thought it would be better
to just refactor using the standard SKB interfaces.  This also means that we
can avoid having to sync Linux to click/net/core/skbuff in the future.