lib: add libdivide
This package is located at [1] with code at [2] and is used to convert fixed
divisions into faster multiplication and shift operations.  The intention
here is to use it in HashContainer, which uses a significant portion of the
CPU on platforms that have a soft divide as part of the modulo operator.
Platforms that have actual operators from this should benefit as well.

The major changes I made to this file were

 * add zlib LICENSE as comment
 * add click glue + friends & fix up #defines
 * cut out all __cplusplus, which is a mess, and every
 * cut out everything not related to u32 division
 * replace 64-bit division with int_divide, for kernel compatibility

The original file is located at [3].

[1] http://libdivide.com/
[2] https://github.com/ridiculousfish/libdivide
[3] https://raw.githubusercontent.com/ridiculousfish/libdivide/master/libdivide.h

Change-Id: I0395693501ef398ad2fe3cb3548372d14200e659

Signed-off-by: Derrick Pallas <pallas@meraki.com>