lib/hashtable: optimize HashContainer::bucket with const division
After @peterhurley noticed that significant time is spent in HashTable, the
modulo operation in HashContainer::bucket was identified as a hotspot.  This
is especially painful on platforms that use soft implementations, e.g. ARM
which implements __aeabi_uidivmod using long division.

We could have converted HashContainer to use power-of-two bucket sizes,
allowing us to replace modulo with a mask.  However, hash tables can suffer
in other ways due to uneven hashing.

There is a significant cost to generate the magic multiplier, the strategy,
and the shift used in the replacement operation.  However the number of
buckets rarely changes and only during expensive operations, so it is worth
converting machine divide instructions into faster machine instructions,
based on libdivide.  Specifically, this will only happen during construction
of the HashContainer and during HashContainer::rehash.

Currently, the libdivide_u32_recover seems to sometimes divide by zero.  It
is not integral to the operation of HashContainer::rehash however, only
appearing in an assert.  The assertion that the magic remainer is the same
as the traditional remainder indicates that this change does work correctly.

YMMV.

Change-Id: I8367a5442d342a786b3fb6b92d326c01d0773e86

Signed-off-by: Derrick Pallas <pallas@meraki.com>