IPFilter: Load and optimize rules in pairs
This patch aims to improve rule compiler optimization runtime.
It does so by changing the order rules are compiled into a program tree.

The original algorithm constructs the program tree by sequentially
appending each of the rules to the end of one single tree,
then running one single optimize operation on that whole tree.

Since the optimization algorithm has a polynomial runtime, this can get
prohibitively slow for large trees.

This patch loads rules two-by-two into small trees, optimizes them, then
combines these trees into a series of smaller, optimized trees. The process
is then repeated until there is only one tree left.

By optimizing each tree as it goes, the new algorithm keeps the tree
size as small as possible, avoiding the polynomial growth in runtime.

The reduction in rule load times varies with the rule set in question,
however we are seeing a reduction in load times of roughly 1/3 on
average. In some cases as much as 97% reduction.

The load time reduction varies quite a bit by rule set, but on average
we are seeing about 1/3 reduction. In some cases the reduction is as
great as 97%, going from 1150ms to 29ms (This rule set had 44 rules, and
45k tree nodes.) These numbers were obtained by running click testies
with various rule sets on server hardware.