Add --cpu=N option to click userlevel
Like linuxmodule cpu parameter, it allows to set up an offset N for thread affinity.
Click's thread I will be pinned to core I+N. If N is not given (just passing --cpu or -a),
 the offset will be 0, which is the old behaviour considering "-a". If the parameter
is not given or N is negative, the thread will have no affinity.

Add a testie to try the new modes

FWY, this is ignored by DPDK as DPDK uses its own mask in EAL arguments.