Fix dumping packets out-of-order when using ListenEtherSwitch
When using the ListenEtherSwitch, which copies all packets it receives
to an extra (the last) port on the switch, packets could appear
out-of-order on that port. This happens when the switch does not know
what port to forward the packet to and decides to broadcast it on all
ports. Take an ARP request for example. As the “dump” port is the last
one, by definition the ARP request will end up at the intended
destination earlier (on one of the real ports). Because of packet
processing, the ARP reply is generated before the request is forwarded
on the last port of the ListenEtherSwitch. On that last port output,
therefore, the reply would be sent before the request.

Simply reversing the broadcast order of the switch ports solves this
problem, as the “dump” port now always outputs the received packet
before any other ports.