Reverse broadcast order
When the broadcast() method is called from within a ListenEtherSwitch,
dumping the packet to the “listen” port last may cause inconsistencies
when e.g. attaching this port to a ToDump.  Because of the packet
processing, this causes e.g. the first ARP reply to appear in the dump
file before the ARP request. Dumping the packet to the listen port
first fixes this issue.