Do not use SELECT_WRITE on push input.
Write events are only needed if input port is pull and run_task already manages adding and removing SELECT_WRITE in that scenario. This commit fixes Click constantly polling selected on push input. Push input does not require polling as it already uses a blocking select in push to setup for writing.