Add IPCounter Element using HashTable_RCU
This element counts ICMP packets hashed on the incoming IP
address. These counts are stored in a HashTable_RCU instance
and acts as an example for safe usage of the structure.

Many of the write/read handlers have the flag included that makes
them execute in a separate linux thread from the rest of click. This
allows for a proof of concept of the multithreadedness of the
structure and element without needing to multithread your entire
click configuration.