ip/ipreassembler: account for header length changes
When storing the first of a series of fragments, IPReassembler copies the IP
header from the fragment to the reassembly buffer.  However, the header of
the new fragment may not be the same length as the original fragment that
created the buffer.  If the new fragment's header is shorter, we will end up
with weirdness preceeding the payload.  If the new fragment's header is
longer, we will overwrite the payload and possibly overrun it.  Both cases
show up as 'bad mem_used' warnings.

Changing header_delta in the p_off=0 case to account for header length
changes (and not just offset changes) fixes the issue.

Found by Waldin Stone, running Defensics.
Signed-off-by: pallas@meraki.com