Added hotwapping support to FromHost and ToHost elements.
FromHost initiates a tap device which is used to pass packets to the kernel
network stack handling. ToHost uses the tap's fd opened be FromHost.
As it was, when sending a new config to click having when the previous had a FromHost
element, would fail as the tap device is busy. Hotswapping support enables to send click
a hotconfig which forces the new FromHost(and ToHost) to take state from the swapped elements,
thus continuing proper packet handling.

When adding hotswap support to FromHost, ToHost must also support hotswap as it searches on initialization
an existing and initialized FromHost element to take it fd.

Signed-off-by: Shlomo Shenzis <shlomo.shenzis@mail.huji.ac.il>