clicktest: parse nested double braces (roughly a no-op)
This change prepares for my next commit in which I'll be making use of
nested double braces. The biggest part of the change is that it
replaces regex based matching with explicit matching. I don't think
it's possible to implement this based on a single somewhat readable
regex.

I also added a unit test which can be run using `clicktest
--self-test-mode`.

Reviewed-by: Yoann Desmouceaux <ydesmouc@meraki.com>