clicktest: add support for integer range expressions
This change adds support for three types of integer range expressions
in %expect and %ignore sections. Here is an example of each type:

{{~ 5 - 15}}    : allow integers from 5 to 15 (inclusive)
{{~ 100 +- 5}}  : allow integers from 95 to 105 (inclusive)
{{~ 100 +- 5%}} : allow integers approximately 5% from 100

These can be used bare as well as inside double-braced regexes (and
similarly inside %expectx and %ignorex sections) such as:
  {{ \d+ buckets? for {{~10 +- 10%}} euros }}

I am hoping this will help developers write tests that are robust
against noise by making it easy to express ranges of permitted values.
Without this, developers need to handcraft equivalent regular
expressions, which is cumbersome and hard to read, e.g.:
{{ 9[5-9]|10[0-5] }}

The current implementation only supports non-negative integers. A
future change could extend this to negative numbers and floats.

I added an algorithm that converts a numeric range to a regex.
Ideally we'd be able to check ranges without using regexes but that
would require a significant rework of clicktest. It was easier to write
the algorithm. The algorithm runs in log time.

Measured performance impact of this and the previous change combined
by running the set of Meraki click tests that take less than 0.2
seconds each (according to parallel's joblog) in serial. I expect
these to have more overhead from the clicktest script than longer
running tests. As follows:

The following alternates between a revert of this commit and this commit:
```
git reset --hard 6a62887e4bd
while :; do
  git rv 6a62887e4bd b23972e0b1d
  git log --oneline HEAD~..HEAD >>before.txt
  (time for t in `perl -ane 'print if $F[3] <0.2' joblog.sorted.txt | awk '{print $11}'`; do ./scripts/clicktest $t; done) 2>&1 1>&2 |grep ^real >>before.txt
  git reset --hard HEAD~~
  git log --oneline HEAD~..HEAD >>after.txt
  (time for t in `perl -ane 'print if $F[3] <0.2' joblog.sorted.txt | awk '{print $11}'`; do ./scripts/clicktest $t; done) 2>&1 1>&2 |grep ^real >>after.txt
done
```

After 328 iterations, I measured an average runtime of 2.633s (before this
commit) and 2.655s (after), with a standard deviation of ~6.5%. This represents
an increase of 0.84%.

Reviewed-by: Yoann Desmouceaux <ydesmouc@meraki.com>